requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2>=2.7.5,<2.8.0
Pillow==8.3.1
flake8==3.9.2
drf-yasg==1.20.0
django-filter==2.4.0
django-cors-headers==3.8.0
django-environ==0.4.5